api_classification_model/requirements.txt
flask
Pillow
torch
transformers
numpy
